requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
autopep8==1.5.6
billiard==3.6.4.0
celery==5.1.0
celery-progress==0.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==0.5.0
Django==3.2
django-celery==3.3.1
django-celery-results==2.0.1
django-debug-toolbar==3.2.1
dnspython==2.1.0
furl==2.1.2
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
kombu==5.1.0
logging-tree==1.9
logx==0.1.1579232358
Markdown==3.3.4
migra==3.0.1621480950
mysqlclient==2.0.3
orderedmultidict==1.0.1
packaging==20.9
pendulum==2.1.2
pip-autoremove==0.9.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-nmap==0.6.4
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
results==0.1.1588569394
schemainspect==3.0.1616029793
six==1.16.0
SQLAlchemy==1.4.17
sqlbag==0.1.1617247075
sqlparse==0.4.1
Sublist3r==1.0
tabulate==0.8.9
toml==0.10.2
Unipath==1.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
